requirements.txt
-i https://pypi.org/simple
appium-python-client==3.2.1
assertpy==1.1
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
cfgv==3.4.0; python_version >= '3.8'
distlib==0.3.8
filelock==3.13.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
identify==2.5.35; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
outcome==1.3.0.post0; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
platformdirs==4.2.0; python_version >= '3.8'
pluggy==1.4.0; python_version >= '3.8'
pre-commit==3.6.2; python_version >= '3.9'
pysocks==1.7.1
pytest==8.0.2; python_version >= '3.8'
python-json-logger==2.0.7; python_version >= '3.6'
pyyaml==6.0.1; python_version >= '3.6'
selenium==4.18.1; python_version >= '3.8'
setuptools==69.1.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sortedcontainers==2.4.0
trio==0.24.0; python_version >= '3.8'
trio-websocket==0.11.1; python_version >= '3.7'
typing-extensions==4.10.0; python_version >= '3.8'
urllib3[socks]==2.2.1; python_version >= '3.8'
virtualenv==20.25.1; python_version >= '3.7'
wsproto==1.2.0; python_full_version >= '3.7.0'
